Fixed %*s length argument type in nxt_sprintf() fmt string.

+++ njs/njs_shell.c
@@ -784,7 +784,7 @@ next:
             }
 
             nxt_sprintf((u_char *) completion, (u_char *) completion + len,
-                        "%*s%V%Z", (int) (p - text), text, suffix);
+                        "%*s%V%Z", p - text, text, suffix);
             return completion;
         }
